Add more tests
